Silence unused parameter warnings for APIs with defined interfaces
Lots of libc interfaces have interfaces which pass arguments which
picolibc doesn't use; silence compiler warnings with (void) param;
statements.

+++ newlib/libc/stdlib/wcstombs.c
@@ -84,12 +84,12 @@ wcstombs (char          *__restrict s,
           bytes = __WCTOMB (buff, *pwcs, &state);
           if (bytes == -1)
             return -1;
-          num_to_copy = (n > bytes ? bytes : (int)n);
+          num_to_copy = (n > (size_t) bytes ? bytes : (int)n);
           for (i = 0; i < num_to_copy; ++i)
             *ptr++ = buff[i];
 
           if (*pwcs == 0x00)
-            return ptr - s - (n >= bytes);
+            return ptr - s - (n >= (size_t) bytes);
           ++pwcs;
           n -= num_to_copy;
         }
